Convert uint64 into uint32 for type size index
This is because we don't support allocations above uint32 in size. So we
check if it's above this and panic.

+++ pkg/store/objectstore/object_store_test.go
@@ -370,6 +370,6 @@ func Test_New_CheckGenericTypeForPointers(t *testing.T) {
 
 func indexForType[T any]() int {
 	t := reflect.TypeFor[T]()
-	size := uint32(t.Size())
+	size := uint64(t.Size())
 	return indexForSize(size)
 }
